+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+## [1.0.0] - 2021-09-01
+### Changed
+- Reworked Icon component to render SVGs icons exported from Figma
+
+### Added
+- Node script to generate icons map and icons collection component for storybook
+
 ## [0.67.3] - 2021-08-16
 ### Added
 - Updated LabelledText colour to grey8
+
 ## [0.67.2] - 2021-08-05
 ### Added
 - Added new border and border color prop to Message component
@@ -80,6 +88,7 @@
 ### Changed
 - Updated gap and styles on Row component
 
+[1.0.0]: https://github.com/marshmallow-insurance/smores-react/compare/v0.67.2...v1.0.0
 [0.67.3]: https://github.com/marshmallow-insurance/smores-react/compare/v0.67.2...v0.67.3
 [0.67.2]: https://github.com/marshmallow-insurance/smores-react/compare/v0.67.1...v0.67.2
 [0.67.1]: https://github.com/marshmallow-insurance/smores-react/compare/v0.67.0...v0.67.1